frc: frc me clock change to 333M from 400M [1/1]
PD#SWPL-52925

Problem:
change me clock

Solution:
change me clock

Verify:
T3

+++ drivers/amlogic/media/frc/frc_reg.c
@@ -114,7 +114,7 @@ EXPORT_SYMBOL(ceil_rx);
 s32 negative_convert(s32 data, u32 fbits)
 {
 	s32 rst = 0;
-	s64 sign_base = 1 << (fbits - 1);
+	s64 sign_base = (s64)1 << (fbits - 1);
 
 	if ((data & sign_base) == sign_base)
 		rst = -((sign_base << 1) - data);
